trainer_service/requirements.txt
anyio==4.13.0
fastapi==0.135.3
httpx==0.28.1
numpy==2.4.3
pydantic==2.12.5
requests==2.33.1
torchmetrics==1.9.0
timm==1.0.26
tqdm==4.67.3
uvicorn==0.43.0